requirements.txt
fastapi
uvicorn
python-dotenv
langchain
langchain-community
langchain-openai
pandas
SQLAlchemy
openpyxl